% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_npy.R
\name{append_npy}
\alias{append_npy}
\title{append_npy}
\usage{
append_npy(
  files1 = NULL,
  files2 = NULL,
  out_filenames = NULL,
  by = "col",
  out_dir = NULL,
  complete_years = T
)
}
\arguments{
\item{files1}{Default = NULL.}

\item{files2}{Default = NULL.}

\item{out_filenames}{Default = NULL. List of filenames of equal length as input data files.}

\item{by}{Default = "col". Append by "col" or "row".}

\item{out_dir}{Default = NULL. Whether diagnostic figures should be produced.}

\item{complete_years}{Default = F. Only used with csv files when col YYY_MM cols exist}
}
\description{
Function to append each file in two lists or folders of csv or npy files by row or col and saving as npy
The function is particular used for wide format data used in models such as Xanthos, Tethys and Demeter.
}
