% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{DATA_VERSION_URL_LIST}
\alias{DATA_VERSION_URL_LIST}
\title{A keyed list containing links to download remote data for each package version in the structure:
list(version_1 = c(data_url_1, data_url_n), #' version_n = c(data_url_1, data_url_n))}
\format{
An object of class \code{list} of length 1.
}
\usage{
DATA_VERSION_URL_LIST
}
\description{
A keyed list containing links to download remote data for each package version in the structure:
list(version_1 = c(data_url_1, data_url_n), #' version_n = c(data_url_1, data_url_n))
}
\keyword{datasets}
