% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.R
\name{delta}
\alias{delta}
\title{delta}
\usage{
delta(
  data = NULL,
  baseline = NULL,
  start_year = 2015,
  save = TRUE,
  diagnostics = TRUE,
  diagnostics_n = 20,
  diagnostics_col = "name",
  filename = NULL,
  folder = NULL,
  output_type = "wide"
)
}
\arguments{
\item{data}{Default = NULL. Input data as R dataframe or CSV.}

\item{baseline}{Default = NULL. Baseline data to apply deltas to.}

\item{start_year}{Default = 2015.}

\item{save}{Default = TRUE. Whether data should be printed as a csv or not.}

\item{diagnostics}{Default = FALSE. Whether diagnostic figures should be produced.}

\item{diagnostics_n}{Default = 20. Number of sub-plots in each diagnostic figure.}

\item{diagnostics_col}{Default = NULL. Which column to plot diagnostics for.}

\item{filename}{Default = NULL}

\item{folder}{Default = NULL}

\item{output_type}{Default = "wide". Output dimensions long or wide.}
}
\description{
Function to calculate deltas from a GCAM base year value
}
\keyword{delta}
