% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_unpack.R
\name{download_file}
\alias{download_file}
\title{Download file from URL path}
\usage{
download_file(url, write_directory, fname)
}
\arguments{
\item{url}{character.  URL to download the target file}

\item{write_directory}{character. Full path to the directory you wish to install the example data to.
Must be write-enabled for the user.}

\item{fname}{character. The file name with extension to save the file as}
}
\description{
Download file and name and save it to a user-specified location
}
\author{
Chris R. Vernon (chris.vernon@pnnl.gov)
}
