% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_unpack.R
\name{download_unpack_zip}
\alias{download_unpack_zip}
\title{Download and unzip zipped file (.zip) to a user-specified location}
\usage{
download_unpack_zip(data_directory, url)
}
\arguments{
\item{data_directory}{character. Full path to the directory you wish to install the example data to.
Must be write-enabled for the user.}

\item{url}{character.  URL to download the target file}
}
\description{
Download and unpack example zipped file with .zip extension to a user-specified location
}
\author{
Chris R. Vernon (chris.vernon@pnnl.gov)
}
