% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_unpack.R
\name{fetch_unpack_data}
\alias{fetch_unpack_data}
\title{Install compressed remote data supplement locally}
\usage{
fetch_unpack_data(data_directory, url_list = DATA_VERSION_URL_LIST)
}
\arguments{
\item{data_directory}{character. Full path to the directory you wish to install the example data to.
Must be write-enabled for the user.}

\item{url_list}{list.  A keyed list containing links to download remote data for each package version in the structure:
list(version_1 = c(data_url_1, data_url_n), #' version_n = c(data_url_1, data_url_n)); default from internal \code{constants.R}}
}
\description{
Download and unpack example data supplement from Zenodo that matches the current installed
distribution.
}
\author{
Chris R. Vernon (chris.vernon@pnnl.gov)
}
