% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_unpack.R
\name{filename_from_url}
\alias{filename_from_url}
\title{Get file name from URL path}
\usage{
filename_from_url(url, file_extension)
}
\arguments{
\item{url}{character.  URL to download the target file}

\item{file_extension}{character. File extension without the dot (e.g., csv)}
}
\description{
Get file name from URL path using the file extension to find the file in the string
}
\author{
Chris R. Vernon (chris.vernon@pnnl.gov)
}
