% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote_read.R
\name{remote_read}
\alias{remote_read}
\title{Read in a file in table format from a remote source.}
\usage{
remote_read(url, ...)
}
\arguments{
\item{url}{character.  URL to the file to import into R, it must be write-enabled for the user.}

\item{...}{additional arguments that may be passed  \code{read.table}}
}
\description{
Currently this function only works with .csv and .txt files.
}
\author{
Kalyn R. Dorheim (kalyn.dorheim@pnnl.gov)
}
