% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth.R
\name{smooth}
\alias{smooth}
\title{smooth}
\usage{
smooth(
  data = NULL,
  window_length = 5,
  window_type = "historical",
  save = TRUE,
  diagnostics = FALSE,
  diagnostics_n = 20,
  diagnostics_col = NULL,
  filename = NULL,
  folder = NULL,
  output_type = "wide"
)
}
\arguments{
\item{data}{Default = NULL. Input data as R dataframe or CSV}

\item{window_length}{Default = 5. Length of smoothing window to use in units of data time period.}

\item{window_type}{Default = "historical". Can be "historical" or "surround" where historical uses
time periods up to the value being smoothed and "surround" uses time period before and after each value being smoothed.}

\item{save}{Default = TRUE. Whether data should be printed as a csv or not.}

\item{diagnostics}{Default = FALSE. Whether diagnostic figures should be produced.}

\item{diagnostics_n}{Default = 20. Number of sub-plots in each diagnostic figure.}

\item{diagnostics_col}{Default = NULL. Which column to plot diagnostics for.}

\item{filename}{Default = NULL}

\item{folder}{Default = NULL}

\item{output_type}{Default = "wide". Output dimensions long or wide.}
}
\value{
Returns a dataframe (and csv file if save = TRUE)
}
\description{
Function to smooth data across different time windows length and position (historical or surrounding)
}
\examples{
library(rpackageutils)
}
\keyword{smoothing}
