% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{version_as_character}
\alias{version_as_character}
\title{Get version as a character dot separated.}
\usage{
version_as_character()
}
\value{
character. For example, c(0, 1, 9) to "0.1.9"
}
\description{
Get the version number as a character dot separated from associated parent package
}
\author{
Chris R. Vernon (chris.vernon@pnnl.gov)
}
