% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xanthos_to_gcam_xml.R
\name{xanthos_to_gcam_xml}
\alias{xanthos_to_gcam_xml}
\title{xanthos_to_gcam_xml}
\source{
gcamdata zchunk_L201.water_resources_constrained
}
\usage{
xanthos_to_gcam_xml(
  xanthos_runoff_csv = NULL,
  gcamdata_folder = NULL,
  out_dir = NULL
)
}
\arguments{
\item{xanthos_runoff_csv}{Default = NULL. Path to xanthos basin runoff .csv outputs file.}

\item{gcamdata_folder}{Default = NULL. Path to gcamdatafolder.}

\item{out_dir}{Default = NULL. Path to folder to save outputs in.}
}
\value{
xml
}
\description{
This function creates xmls for GCAM from raw xanthos outputs
a given dataframe. Returns a long dataframe with lat lon and parameter from the netcdf chosen.
}
